Fixed issue in get question by Id

+++ src/services/question.js
@@ -10,8 +10,9 @@ export const retrieveQuestion = async (question_id) => {
     return await getQuestionById(question_id);
 }
 
-export const deleteQuestion = async (question_id) => {
+export const deleteQuestion = async (question_id, user) => {
     const question = await findQuestion({ _id: question_id })
     if (!question) return { status: 400, message: 'Question does not exist to remove' }
+    if (question.creator_lock && question.creator !== user._id) return { status: 403, message: 'You are not authorized to delete this question' }
     return await deleteAQuestion({ _id: question_id })
 }
